+++ pet-foods-be/src/main/java/com/petfoods/be/dto/FoodDto.java
@@ -7,8 +7,8 @@ public class FoodDto {
 	private Long id;
 	private String marca; 
 	private String descrizione;
-	private Long prezzo;
-	private Long peso;
+	private Double prezzo;
+	private Double peso;
 	private String unitaDiMisura;
 	private List<ComponentiAnaliticiDto> componentiAnalitici;
 	
@@ -30,16 +30,16 @@ public class FoodDto {
 	public void setDescrizione(String descrizione) {
 		this.descrizione = descrizione;
 	}
-	public Long getPrezzo() {
+	public Double getPrezzo() {
 		return prezzo;
 	}
-	public void setPrezzo(Long prezzo) {
+	public void setPrezzo(Double prezzo) {
 		this.prezzo = prezzo;
 	}
-	public Long getPeso() {
+	public Double getPeso() {
 		return peso;
 	}
-	public void setPeso(Long peso) {
+	public void setPeso(Double peso) {
 		this.peso = peso;
 	}
 	public String getUnitaDiMisura() {
